+++ packages/core/icon-file-type/src/components/__tests__/unit/server-side-hydrate.tsx
@@ -8,10 +8,10 @@ jest.spyOn(global.console, 'error').mockImplementation(() => {});
 afterEach(() => {
   jest.resetAllMocks();
 });
-// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
-test.skip('should ssr then hydrate icon-file-interface correctly', async () => {
+
+test('should ssr then hydrate icon-file-interface correctly', async () => {
   const [example] = await getExamplesFor('icon-file-type');
-  const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
+  const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
 
   const elem = document.createElement('div');
   elem.innerHTML = await ssr(example.filePath);