+++ packages/core/icon-priority/src/components/__tests__/unit/server-side-hydrate.tsx
@@ -6,16 +6,13 @@ import waitForExpect from 'wait-for-expect';
 
 jest.spyOn(global.console, 'error').mockImplementation(() => {});
 
-beforeEach(() => {
-  jest.setTimeout(10000);
-});
 afterEach(() => {
   jest.resetAllMocks();
 });
-// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
-test.skip('should ssr then hydrate icon-priority correctly', async () => {
+
+test('should ssr then hydrate icon-priority correctly', async () => {
   const [example] = await getExamplesFor('icon-priority');
-  const Example = await require(example.filePath).default;
+  const Example = require(example.filePath).default;
 
   const elem = document.createElement('div');
   elem.innerHTML = await ssr(example.filePath);