+++ packages/core/popper/src/__tests__/unit/server-side-hydrate.tsx
@@ -4,10 +4,6 @@ import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
 import { ssr } from '@atlaskit/ssr';
 import waitForExpect from 'wait-for-expect';
 
-beforeEach(() => {
-  jest.setTimeout(10000);
-});
-
 jest.mock('popper.js', () => {
   const PopperJS = require.requireActual('popper.js');
 
@@ -29,10 +25,10 @@ jest.spyOn(global.console, 'error').mockImplementation(() => {});
 afterEach(() => {
   jest.resetAllMocks();
 });
-// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
-test.skip('should ssr then hydrate popper correctly', async () => {
+
+test('should ssr then hydrate popper correctly', async () => {
   const [example] = await getExamplesFor('popper');
-  const Example = await require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
+  const Example = require(example.filePath).default; // eslint-disable-line import/no-dynamic-require
 
   const elem = document.createElement('div');
   elem.innerHTML = await ssr(example.filePath);