+++ packages/core/popup/examples/12-multiple-popups.tsx
@@ -1,14 +1,18 @@
 /** @jsx jsx */
-import { FC, useCallback, useState } from 'react';
+import { Component, FC, useCallback, useState } from 'react';
 import Button from '@atlaskit/button';
 import { jsx } from '@emotion/core';
-import Popup from '../src';
+import Popup, { ContentProps } from '../src';
 
 const contentCSS = {
   margin: '1rem',
 };
 
-const PopupContent: FC<{}> = () => <div css={contentCSS}>content</div>;
+class PopupContent extends Component<ContentProps> {
+  render() {
+    return <div css={contentCSS}>content</div>;
+  }
+}
 
 type PopupExampleProps = {
   name: string;