+++ packages/media/media-client/src/__tests__/unit/media-store.spec.ts
@@ -270,11 +270,15 @@ describe('MediaStore', () => {
         return mediaStore.getFile(fileId, params).then(response => {
           expect(response).toEqual({ data });
           expect(fetchMock.lastUrl()).toEqual(
-            `${baseUrl}/file/${fileId}?client=${clientId}&collection=${collectionName}&token=${token}`,
+            `${baseUrl}/file/${fileId}?collection=${collectionName}`,
           );
           expect(fetchMock.lastOptions()).toEqual(
             expect.objectContaining({
               method: 'GET',
+              headers: {
+                'X-Client-Id': clientId,
+                Authorization: `Bearer ${token}`,
+              },
             }),
           );
           expect(authProvider).toHaveBeenCalledWith({ collectionName });
@@ -367,12 +371,14 @@ describe('MediaStore', () => {
         return mediaStore.getCollection(collectionName).then(response => {
           expect(response).toEqual({ data });
           expect(fetchMock.lastUrl()).toEqual(
-            `${baseUrl}/collection/${collectionName}?client=${clientId}&token=${token}`,
+            `${baseUrl}/collection/${collectionName}`,
           );
           expect(fetchMock.lastOptions()).toEqual({
             method: 'GET',
             headers: {
               Accept: 'application/json',
+              'X-Client-Id': clientId,
+              Authorization: `Bearer ${token}`,
             },
           });
           expect(authProvider).toHaveBeenCalledWith({ collectionName });
@@ -405,12 +411,14 @@ describe('MediaStore', () => {
           .then(response => {
             expect(response).toEqual({ data });
             expect(fetchMock.lastUrl()).toEqual(
-              `${baseUrl}/collection/some-collection-name/items?client=${clientId}&details=full&inclusiveStartKey=some-inclusive-start-key&limit=10&sortDirection=desc&token=${token}`,
+              `${baseUrl}/collection/some-collection-name/items?details=full&inclusiveStartKey=some-inclusive-start-key&limit=10&sortDirection=desc`,
             );
             expect(fetchMock.lastOptions()).toEqual({
               method: 'GET',
               headers: {
                 Accept: 'application/json',
+                'X-Client-Id': clientId,
+                Authorization: `Bearer ${token}`,
               },
             });
             expect(authProvider).toHaveBeenCalledWith({ collectionName });
@@ -537,11 +545,11 @@ describe('MediaStore', () => {
       });
 
       it('should fail if error status is returned', () => {
-        const error = {
+        const errorBody = {
           error: 'something wrong',
         };
         fetchMock.mock(`begin:${baseUrl}/upload/createWithFiles`, {
-          body: error,
+          body: errorBody,
           status: 403,
         });
 
@@ -552,9 +560,10 @@ describe('MediaStore', () => {
           result => {
             expect(result).not.toBeDefined();
           },
-          async (response: Response) => {
-            const reason = await response.json();
-            expect(reason).toEqual(error);
+          async error => {
+            expect(error.message).toEqual(
+              'network error 403: {"error":"something wrong"}',
+            );
           },
         );
       });
@@ -615,7 +624,7 @@ describe('MediaStore', () => {
 
         const image = await mediaStore.getImage('123');
         expect(fetchMock.lastUrl()).toEqual(
-          `${baseUrl}/file/123/image?allowAnimated=true&client=some-client-id&max-age=${FILE_CACHE_MAX_AGE}&mode=crop&token=some-token`,
+          `${baseUrl}/file/123/image?allowAnimated=true&max-age=${FILE_CACHE_MAX_AGE}&mode=crop`,
         );
         expect(image).toBeInstanceOf(Blob);
       });
@@ -634,7 +643,7 @@ describe('MediaStore', () => {
           upscale: true,
         });
         expect(fetchMock.lastUrl()).toEqual(
-          `${baseUrl}/file/123/image?allowAnimated=true&client=some-client-id&max-age=${FILE_CACHE_MAX_AGE}&mode=full-fit&token=some-token&upscale=true&version=2`,
+          `${baseUrl}/file/123/image?allowAnimated=true&max-age=${FILE_CACHE_MAX_AGE}&mode=full-fit&upscale=true&version=2`,
         );
       });
 
@@ -657,7 +666,7 @@ describe('MediaStore', () => {
           true,
         );
         expect(fetchMock.lastUrl()).toEqual(
-          `${baseUrl}/file/123/image?allowAnimated=true&client=some-client-id&height=4096&max-age=${FILE_CACHE_MAX_AGE}&mode=full-fit&token=some-token&upscale=true&version=2&width=4096`,
+          `${baseUrl}/file/123/image?allowAnimated=true&height=4096&max-age=${FILE_CACHE_MAX_AGE}&mode=full-fit&upscale=true&version=2&width=4096`,
         );
       });
 
@@ -682,7 +691,7 @@ describe('MediaStore', () => {
           true,
         );
         expect(fetchMock.lastUrl()).toEqual(
-          `${baseUrl}/file/123/image?allowAnimated=true&client=some-client-id&height=4096&max-age=${FILE_CACHE_MAX_AGE}&mode=crop&token=some-token&upscale=true&version=2&width=4096`,
+          `${baseUrl}/file/123/image?allowAnimated=true&height=4096&max-age=${FILE_CACHE_MAX_AGE}&mode=crop&upscale=true&version=2&width=4096`,
         );
       });
 
@@ -799,7 +808,7 @@ describe('MediaStore', () => {
 
         const image = await mediaStore.getImageMetadata('123');
         expect(fetchMock.lastUrl()).toEqual(
-          `${baseUrl}/file/123/image/metadata?client=some-client-id&token=some-token`,
+          `${baseUrl}/file/123/image/metadata`,
         );
         expect(image).toEqual(data);
       });
@@ -817,7 +826,7 @@ describe('MediaStore', () => {
           collection: 'my-collection',
         });
         expect(fetchMock.lastUrl()).toEqual(
-          `${baseUrl}/file/123/image/metadata?client=some-client-id&collection=my-collection&token=some-token`,
+          `${baseUrl}/file/123/image/metadata?collection=my-collection`,
         );
       });
     });