+++ packages/media/media-picker/src/popup/middleware/analyticsHandlers/editRemoteImageHandler.ts
@@ -4,10 +4,7 @@ import { buttonClickPayload, HandlerResult } from '.';
 
 export default (action: Action): HandlerResult => {
   if (isEditRemoteImageAction(action)) {
-    const {
-      collectionName,
-      item: { id = undefined, name = undefined } = {},
-    } = action;
+    const { collectionName, item: { id = undefined } = {} } = action;
     return [
       {
         ...buttonClickPayload,
@@ -15,7 +12,6 @@ export default (action: Action): HandlerResult => {
         attributes: {
           collectionName,
           fileId: id,
-          fileName: name,
         },
       },
     ];