+++ packages/media/media-picker/src/service/uploadServiceImpl.ts
@@ -188,7 +188,7 @@ export class UploadServiceImpl implements UploadService {
             }
 
             if (state.status === 'processing') {
-              subscription.unsubscribe();
+              setTimeout(() => subscription.unsubscribe(), 0);
               if (shouldCopyFileToRecents) {
                 mediaClient.emit('file-added', state);
                 globalMediaEventEmitter.emit('file-added', state);