+++ packages/navigation/atlassian-switcher/examples/21-themed-switcher.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Button from '@atlaskit/button';
 import Drawer from '@atlaskit/drawer';
-import { mockEndpoints, REQUEST_FAST } from './helpers/mock-endpoints';
+import { mockEndpoints } from '@atlaskit/atlassian-switcher-test-utils';
 import { withIntlProvider } from './helpers';
 import AtlassianSwitcher from '../src';
 class JiraSwitcherExample extends React.Component {
@@ -10,7 +10,7 @@ class JiraSwitcherExample extends React.Component {
   };
 
   openDrawer = () => {
-    mockEndpoints('jira', originalMockData => originalMockData, REQUEST_FAST);
+    mockEndpoints('jira');
     this.setState({
       isDrawerOpen: true,
     });