+++ packages/navigation/atlassian-switcher/src/components/loaders.tsx
@@ -1,16 +1,16 @@
 import Loadable from 'react-loadable';
 
 export const loadAtlassianSwitcher = () =>
-  import(/* webpackChunkName: "@atlaskit/atlassian-switcher/async-chunk/atlassian-switcher" */ './atlassian-switcher');
+  import(/* webpackChunkName: "@ak-switcher-chunk-atlassian-switcher" */ './atlassian-switcher');
 
 export const loadJiraSwitcher = () =>
-  import(/* webpackChunkName: "@atlaskit/atlassian-switcher/async-chunk/jira-switcher" */ './jira-switcher');
+  import(/* webpackChunkName: "@ak-switcher-chunk-jira-switcher" */ './jira-switcher');
 
 export const loadConfluenceSwitcher = () =>
-  import(/* webpackChunkName: "@atlaskit/atlassian-switcher/async-chunk/confluence-switcher" */ './confluence-switcher');
+  import(/* webpackChunkName: "@ak-switcher-chunk-confluence-switcher" */ './confluence-switcher');
 
 export const loadGenericSwitcher = () =>
-  import(/* webpackChunkName: "@atlaskit/atlassian-switcher/async-chunk/generic-switcher" */ './generic-switcher');
+  import(/* webpackChunkName: "@ak-switcher-chunk-generic-switcher" */ './generic-switcher');
 
 export const AtlassianSwitcherLoader = Loadable({
   loader: loadAtlassianSwitcher,