+++ packages/navigation/atlassian-switcher/src/providers/recommendations/index.ts
@@ -4,13 +4,17 @@ import {
   RecommendationsFeatureFlags,
 } from '../../types';
 
-import { jswOgExpandsExperiment } from './experiments';
+import {
+  jswOgExpandsExperiment,
+  productStoreRecommendation,
+} from './experiments';
 
 function baseRecommendation(): RecommendationItem[] {
   return [
     { productKey: ProductKey.JIRA_SOFTWARE },
     { productKey: ProductKey.CONFLUENCE },
     { productKey: ProductKey.JIRA_SERVICE_DESK },
+    { productKey: ProductKey.OPSGENIE },
   ];
 }
 
@@ -27,6 +31,10 @@ export function resolveRecommendations(
     )
   ) {
     return jswOgExpandsExperiment.recommendations();
+  }
+
+  if (featureFlags.isDiscoverSectionEnabled) {
+    return productStoreRecommendation();
   }
 
   return baseRecommendation();