+++ packages/editor/renderer/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/renderer
 
+## 51.1.2
+
+- Updated dependencies [4778521db3](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4778521db3):
+  - @atlaskit/navigation-next@7.0.0
+
 ## 51.1.1
 
 ### Patch Changes