+++ packages/navigation/atlassian-switcher/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @atlaskit/atlassian-switcher
 
+## 4.6.4
+
+### Patch Changes
+
+- [patch][6bc87c7501](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6bc87c7501):
+
+  Split mockEndpoints into a separate package
+
+## 4.6.3
+
+### Patch Changes
+
+- [patch][f07586ba0e](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f07586ba0e):
+
+  Hide manage list link when custom links are disabled
+
+## 4.6.2
+
+### Patch Changes
+
+- [patch][14e55631c6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/14e55631c6):
+
+  Add groupIndex to switcher item analytics
+
+## 4.6.1
+
+### Patch Changes
+
+- [patch][e6f7f7747f](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e6f7f7747f):
+
+  Add Opsgenie to default list of product recommendations
+
+## 4.6.0
+
+### Minor Changes
+
+- [minor][cd96adc5e7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/cd96adc5e7):
+
+  Add Try lozenge to first item in Discover section
+
+## 4.5.4
+
+### Patch Changes
+
+- [patch][d904d4df82](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d904d4df82):
+
+  Use kebab case for switcher chunk names
+
+## 4.5.3
+
+### Patch Changes
+
+- [patch][03e00bd3d7](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/03e00bd3d7):
+
+  Fix discover/x-sell work for Bitbucket cloud
+
+## 4.5.2
+
+### Patch Changes
+
+- [patch][be38f90454](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/be38f90454):
+
+  Added support for additional languages: Thai, Turkish, Ukrainia and Vietnamese
+
+## 4.5.1
+
+### Patch Changes
+
+- [patch][165f1cd0ad](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/165f1cd0ad):
+
+  Refactor codes related to experimental Discover section in Atlassian Switcher.
+
+## 4.5.0
+
+### Minor Changes
+
+- [minor][331053f02c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/331053f02c):
+
+  Add experimental discover section to Atlassian Switcher.
+
 ## 4.4.0
 
 ### Minor Changes